+++ shared/subgraph.cu
@@ -6,7 +6,7 @@
 
 
 template <class E>
-Subgraph<E>::Subgraph(uint num_nodes, uint num_edges)
+Subgraph<E>::Subgraph(uint num_nodes, u_int64_t num_edges)
 {
 	cudaProfilerStart();
 	cudaError_t error;
@@ -33,11 +33,11 @@ Subgraph<E>::Subgraph(uint num_nodes, uint num_edges)
 	this->num_edges = num_edges;
 	
 	gpuErrorcheck(cudaMallocHost(&activeNodes, num_nodes * sizeof(uint)));
-	gpuErrorcheck(cudaMallocHost(&activeNodesPointer, (num_nodes+1) * sizeof(uint)));
+	gpuErrorcheck(cudaMallocHost(&activeNodesPointer, (num_nodes+1) * sizeof(u_int64_t)));
 	gpuErrorcheck(cudaMallocHost(&activeEdgeList, num_edges * sizeof(E)));
 	
 	gpuErrorcheck(cudaMalloc(&d_activeNodes, num_nodes * sizeof(unsigned int)));
-	gpuErrorcheck(cudaMalloc(&d_activeNodesPointer, (num_nodes+1) * sizeof(unsigned int)));
+	gpuErrorcheck(cudaMalloc(&d_activeNodesPointer, (num_nodes+1) * sizeof(u_int64_t)));
 	gpuErrorcheck(cudaMalloc(&d_activeEdgeList, (max_partition_size) * sizeof(E)));
 }
